Renamed LoginRepository to AuthenticationRepository and added class to refresh token. - Kyle Barry (u19232510)

+++ mobile/app/src/main/java/com/bdpsolutions/highfive/subsystems/login/model/dataclass/AuthToken.kt
@@ -12,6 +12,6 @@ data class AuthToken(
     @PrimaryKey val uid: Int,
     @ColumnInfo(name = "auth_token") val authToken: String?,
     @ColumnInfo(name = "refresh_token") val refreshToken: String?,
-    @ColumnInfo(name = "auth_expires") val authExpires: Int?,
-    @ColumnInfo(name = "refresh_expires") val refreshExpires: Int?
+    @ColumnInfo(name = "auth_expires") val authExpires: Long?,
+    @ColumnInfo(name = "refresh_expires") val refreshExpires: Long?
 )
